experiment with calling classes
I added "div.content" to change the font size and width of the post content on the home page. I want to see if trying this new way to call class, "content", will work on your page. I tried it on my page and Jekyll failed to build my page. The css selectors I used were element and class.